feat: introduce dog wag tail loading component.
Introduce a new 'Dog Wag Tail' component for a playful loading screen animation. The animation features a dog wagging its tail and moving its body to the sides.

Adjusted the cards layout by adding a margin-bottom to create space to the elements. This change prevents the cards from being nearly hidden when scrolled to the end of the page.

Author: Wellesley Mussolini